requirements.txt
Flask>=3.1.1
flask_login>=0.6.3
sqlalchemy>=2.0.43
python-dotenv>=1.0.0
geopy>=2.4.1
bcrypt>=4.3.0
psycopg2>=2.9.10
requests>=2.32.5